pyproject.toml
[project]
name = "arion-stream-architecture"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "python-dotenv>=1.0.1",
    "python-on-whales>=0.75.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.3",
]
